Add .gitignore
We don't want to track the sites/default/settings.local.php file. The
whole point of having it is to allow us to maintain different settings
in different environments. If we committed it, then it would get pulled
into all of the environments.

In reality, we would also not want to keep the settings.php in our
repository because we wouldn't want to risk exposing sensitive
information. However, for the sake of keeping things complete and
transparent for the class we chose not to use a .gitignore prior to now.