Capture all the things!
Note: As with all of the updates in section 18, this is being done in
the development environment.

This commit captures updating Drupal to 8.4.0.

Note 2: This update DOES require the database to be updated! After you
pull in these code changes make sure to run a database update.

Somewhere in Drupal 8.3.6 or 8.3.7 there was a regression that was
introduced that caused the site name and slogan that we had overridden
in our local settings file to not display when we were signed in. If we
were signed out, both showed the overridden values. If we were signed in
we saw the values that are in configuration. However, the <title> tag
did correctly pick up our overridden site name. I haven't found any
mention of this issue in the release notes. Nonetheless, from my own
observation I could see that it was working in Drupal 8.3.5; it wasn't
working in Drupal 8.3.7; and it is now working again in Drupal 8.4.0. I
had already checked whether rebuilding the cache file fixed the problem
and saw that it didn't.